+++ src/Board.tsx
@@ -1,13 +1,11 @@
 import styled from 'styled-components';
 import _ from 'underscore';
-import Hand from './ui/Hand';
 import HiddenHand from './ui/HiddenHand';
 import Stable, { StableHandle } from './ui/Stable';
 import PlayerField, { PlayerFieldHandle } from './ui/PlayerField';
-import useSound from 'use-sound';
 import { motion, AnimateSharedLayout, AnimatePresence } from "framer-motion";
 // game
-import { UnstableUnicornsGame, Ctx, _findOpenScenesWithProtagonist, Instruction, Scene, canPlayCard, _findInProgressScenesWithProtagonist, _findInstruction } from './game/game';
+import { UnstableUnicornsGame, Ctx, _findInProgressScenesWithProtagonist, _findOpenScenesWithProtagonist, Instruction, Scene } from './game/game';
 // assets
 import BG from './assets/ui/board-background.jpg';
 import DrawPile from './ui/DrawPile';
@@ -15,34 +13,21 @@ import Nursery from './ui/Nursery';
 import DiscardPile from './ui/DiscardPile';
 import { CardID } from './game/card';
 import { useContext, useEffect, useRef, useState } from 'react';
-import { findUITargets, HoverTarget } from './BoardUtil';
-import RainbowArrow from './ui/RainbowArrow';
+import { CardInteraction, HoverTarget } from './BoardUtil';
 import EndTurnButton from './ui/button/EndTurnButton';
 import { PlayerID } from './game/player';
 import { BoardState, getBoardState } from './BoardStateManager';
 import type { Moves } from './game/types';
 import GameLabel from './ui/GameLabel';
-import NeighLabel, { NeighLabelRole } from './ui/NeighLabel';
-import CardPopupSingleAction from './ui/CardPopupSingleAction';
-import { AddFromDiscardPileToHandTarget, BringToStableTarget, DiscardTarget, ReviveTarget, SearchTarget } from './game/do';
-import InfoLabel from './ui/InfoLabel';
-import Finder from './ui/Finder';
+import { SearchTarget } from './game/do';
 import BoardGameBegin from './BoardGameBegin';
 import React from 'react';
 import { LanguageContext } from './LanguageContextProvider';
-
-const YourTurnSound = require('./assets/sound/ALERT_YourTurn_0v2.ogg');
-const DrawCardSound = require('./assets/sound/draw_card_and_add_to_hand_1.ogg');
-const NeighSuccessSound = require('./assets/sound/tavern_crowd_play_reaction_very_positive_5.ogg');
-const NeighFailureSound = require('./assets/sound/tavern_crowd_play_reaction_quite_positive_2.ogg');
-const EndTurnButtonSound = require('./assets/sound/bar_button_A_press.ogg');
-const HubMouseOverSound = require('./assets/sound/Hub_Mouseover.ogg');
-const FriendlyChallengeSound = require('./assets/sound/friendly_challenge.ogg');
-const DestroyedSound = require('./assets/sound/Arrow_Targeted_Explode_01.ogg');
-const DiscardSound = require('./assets/sound/play_card_from_hand_1.ogg');
-const SacrificeSound = require('./assets/sound/FeignDeath_trigger_1.ogg');
-const StealSound = require('./assets/sound/stealth_on.ogg');
-
+import { useSoundEffects } from './hooks/useSoundEffects';
+import NeighPanel from './components/NeighPanel';
+import InfoPanel from './components/InfoPanel';
+import OverlayManager from './components/OverlayManager';
+import StableSection from './components/StableSection';
 
 type Props = {
     G: UnstableUnicornsGame;
@@ -52,129 +37,10 @@ type Props = {
     isActive: boolean;
 }
 
-type CardInteraction = {
-    key: "card_to_card",
-    info: {
-        sourceCardID: CardID,
-        instructionID: string,
-        currentMousePosition: { x: number, y: number },
-        startingMousePosition: { x: number, y: number },
-        targets: { cardID: CardID }[]
-    }
-} | {
-    key: "card_to_player",
-    info: {
-        sourceCardID: CardID,
-        instructionID: string,
-        currentMousePosition: { x: number, y: number },
-        startingMousePosition: { x: number, y: number },
-        targets: { playerID: PlayerID }[]
-    }
-} | {
-    key: "click_on_other_stable_card",
-    info: {
-        instructionID: string,
-        targets: { cardID: CardID }[],
-    }
-} | {
-    key: "play_upgradeDowngradeCardFromHand__choose_target",
-    info: {
-        instructionID: "____________",
-        currentMousePosition: { x: number, y: number },
-        startingMousePosition: { x: number, y: number },
-        cardID: CardID,
-    }
-}
-
 const Board = (props: Props) => {
     const { G, ctx, playerID, moves } = props;
 
-    const [playYourTurnAlert] = useSound(YourTurnSound, {
-        volume: 0.3,
-    });
-
-    const [playDrawCardSound] = useSound(DrawCardSound, {
-        volume: 0.3,
-    });
-
-    const [playNeighSuccessSound] = useSound(NeighSuccessSound, {
-        volume: 0.2,
-    });
-
-    const [playNeighFailureSound] = useSound(NeighFailureSound, {
-        volume: 0.2,
-    });
-
-    const [playEndTurnButtonSound] = useSound(EndTurnButtonSound, {
-        volume: 0.2,
-    });
-
-    const [playHubMouseOverSound] = useSound(HubMouseOverSound, {
-        volume: 0.2,
-    });
-
-    const [playFriendlyChallengeSound] = useSound(FriendlyChallengeSound, {
-        volume: 0.2,
-    });
-
-    const [playDestroyedSound] = useSound(DestroyedSound, {
-        volume: 0.2,
-    });
-
-    const [playDiscardSound] = useSound(DiscardSound, {
-        volume: 0.2,
-    });
-
-    const [playSacrificeSound] = useSound(SacrificeSound, {
-        volume: 0.2,
-    });
-
-    const [playStealSound] = useSound(StealSound, {
-        volume: 0.2,
-    });
-
-    useEffect(() => {
-        if (ctx.currentPlayer === playerID) {
-            playYourTurnAlert();
-        }
-    }, [ctx.currentPlayer]);
-
-    useEffect(() => {
-        if (G.lastNeighResult) {
-            if (G.lastNeighResult.result === "cardWasPlayed") {
-                playNeighSuccessSound();
-            } else {
-                playNeighFailureSound();
-            }
-        }
-    }, [G.lastNeighResult?.id]);
-
-    useEffect(() => {
-        if (G.neighDiscussion?.rounds.length !== undefined && G.neighDiscussion?.rounds.length > 0) {
-            playFriendlyChallengeSound();
-        }
-    }, [G.neighDiscussion?.rounds.length])
-
-    useEffect(() => {
-        if (G.uiExecuteDo) {
-            if (G.uiExecuteDo.do.key === "destroy") {
-                playDestroyedSound();
-            }
-
-            if (G.uiExecuteDo.do.key === "sacrifice") {
-                playSacrificeSound();
-            }
-
-            if (G.uiExecuteDo.do.key === "discard") {
-                playDiscardSound();
-            }
-
-            if (G.uiExecuteDo.do.key === "steal") {
-                playStealSound();
-            }
-        }
-    }, [G.uiExecuteDo?.id]);
-
+    const { playDrawCardSound, playEndTurnButtonSound, playHubMouseOverSound, playExecuteDoSound } = useSoundEffects(G, ctx, playerID);
 
     const [showDeckFinder, setShowDeckFinder] = useState<SearchTarget[] | undefined>(undefined);
     const [showPlayerHand, setShowPlayerHand] = useState<PlayerID | undefined>(undefined);
@@ -232,7 +98,7 @@ const Board = (props: Props) => {
     const boardStates = getBoardState(G, ctx, playerID);
     if (boardStates.find(s => s.type === "destroy__click_on_card_in_stable" || s.type === "sacrifice__clickOnCardInStable")) {
         const boardState = boardStates.find(s => s.type === "destroy__click_on_card_in_stable" || s.type === "sacrifice__clickOnCardInStable")!;
-        // only update if the card interaction is different 
+        // only update if the card interaction is different
         // prevents the client from rendering indefinitely
         if (cardInteraction?.info?.instructionID !== boardState.info!.instructionID) {
             setCardInteraction({
@@ -244,46 +110,7 @@ const Board = (props: Props) => {
         }
     }
 
-    // Card to Card Hover Interaction Online
     const [C2CArrow, setC2CArrow] = useState<{ fromX: number, fromY: number, toX: number, toY: number } | undefined>(undefined);
-
-    useEffect(() => {
-        if (G.uiCardToCard?.protagonist === playerID) {
-            return;
-        }
-
-        if (G.uiCardToCard !== undefined && G.uiCardToCard.sourceCardID && G.uiCardToCard.targetCardID) {
-            if (playerFieldRef.current && stableRef.current) {
-                const refSource = playerFieldRef.current.getStableItemRef(G.uiCardToCard.sourceCardID);
-                let targetSource;
-                if (G.stable[playerID].find(i => i === G.uiCardToCard!.targetCardID)) {
-                    targetSource = stableRef.current.getStableItemRef(G.uiCardToCard.targetCardID);
-                } else {
-                    targetSource = playerFieldRef.current.getStableItemRef(G.uiCardToCard.targetCardID);
-                }
-
-                if (targetSource?.current && refSource?.current) {
-                    const coord = refSource.current!.getBoundingClientRect()!;
-                    const from = {
-                        x: coord.left + coord.width / 2.0,
-                        y: coord.top + coord.height / 2.0,
-                    };
-                    const coord2 = targetSource.current!.getBoundingClientRect()!;
-                    const to = {
-                        x: coord2.left + coord2.width / 2.0,
-                        y: coord2.top + coord2.height / 2.0,
-                    };
-
-
-                    setC2CArrow({
-                        fromX: from.x, fromY: from.y, toX: to.x, toY: to.y
-                    });
-                }
-            }
-        }  else {
-            setC2CArrow(undefined);
-        }
-    }, [G.uiCardToCard?.id]);
 
 
     if (ctx.phase === "pregame") {
@@ -310,70 +137,27 @@ const Board = (props: Props) => {
                     Englisch
                 </div>
 
-                
-                {showDeckFinder &&
-                    <Finder
-                        cards={showDeckFinder.map(s => G.deck[s.cardID])}
-                        showBackButton={false}
-                        onBackClick={() => 0}
-                        onCardClick={cardID => {
-                            const boardState = boardStates.find(o => o.type === "search__single_action_popup")!;
-                            const { instruction } = _findInstruction(G, boardState.info!.instructionID)!;
-                            moves.executeDo(instruction.id, {
-                                protagonist: playerID,
-                                cardID
-                            });
-                            setShowDeckFinder(undefined);
-                            setCardInteraction(undefined);
-                        }} />
-                }
-                {showDiscardFinder &&
-                    <Finder
-                        cards={showDiscardFinder.map(c => G.deck[c.cardID])}
-                        onBackClick={() => setShowDiscardFinder(undefined)}
-                        onCardClick={cardID => {
-                            let state = boardStates.find(s => s.type === "revive" || s.type === "addFromDiscardPileToHand__single_action_popup");
-                            if (state) {
-                                moves.executeDo(state.info?.instructionID, {
-                                    protagonist: playerID, cardID
-                                });
 
-                                setShowDiscardFinder(undefined);
-                            }
-                        }} />
-                }
-                {showNurseryFinder &&
-                    <Finder cards={G.nursery.map(c => G.deck[c])} onBackClick={() => setShowNurseryFinder(false)} onCardClick={cardID => {
-                        let state = boardStates.find(s => s.type === "reviveFromNursery");
-                        if (state) {
-                            moves.executeDo(state.info?.instructionID, {
-                                protagonist: playerID, cardID
-                            });
+                <OverlayManager
+                    G={G}
+                    boardStates={boardStates}
+                    moves={moves}
+                    playerID={playerID}
+                    cardInteraction={cardInteraction}
+                    setCardInteraction={setCardInteraction}
+                    showDeckFinder={showDeckFinder}
+                    setShowDeckFinder={setShowDeckFinder}
+                    showDiscardFinder={showDiscardFinder}
+                    setShowDiscardFinder={setShowDiscardFinder}
+                    showNurseryFinder={showNurseryFinder}
+                    setShowNurseryFinder={setShowNurseryFinder}
+                    showPlayerHand={showPlayerHand}
+                    setShowPlayerHand={setShowPlayerHand}
+                    showBlatantThievery={showBlatantThievery}
+                    setShowBlatantThievery={setShowBlatantThievery}
+                    C2CArrow={C2CArrow}
+                />
 
-                            setShowNurseryFinder(false);
-                        }
-                    }} />
-                }
-                {showPlayerHand !== undefined &&
-                    <Finder cards={G.hand[showPlayerHand].map(c => G.deck[c])} onBackClick={() => setShowPlayerHand(undefined)} showBackButton={true} onCardClick={cardID => {
-
-                    }} hide={G.playerEffects[showPlayerHand].find(o => o.effect.key === "your_hand_is_visible") === undefined} />
-                }
-                {showBlatantThievery !== undefined &&
-                    <Finder cards={G.hand[showBlatantThievery].map(c => G.deck[c])} onBackClick={() => undefined} showBackButton={false} onCardClick={cardID => {
-                        const handIndex = G.hand[showBlatantThievery].findIndex(s => s === cardID);
-                        moves.executeDo(cardInteraction?.info.instructionID, { protagonist: playerID, handIndex, from: showBlatantThievery });
-                        setShowBlatantThievery(undefined);
-                        setCardInteraction(undefined);
-                    }}
-                        title="Click on card to add it to your hand." />
-                }
-                {(cardInteraction?.key === "card_to_card" || cardInteraction?.key === "card_to_player" || cardInteraction?.key === "play_upgradeDowngradeCardFromHand__choose_target") &&
-                    <RainbowArrow from={{ x: cardInteraction.info.startingMousePosition.x, y: cardInteraction.info.startingMousePosition.y }} to={{ x: cardInteraction.info.currentMousePosition.x, y: cardInteraction.info.currentMousePosition.y }} />
-                }
-                {C2CArrow !== undefined &&
-                    <RainbowArrow from={{ x: C2CArrow!.fromX, y: C2CArrow!.fromY }} to={{ x: C2CArrow!.toX, y: C2CArrow!.toY }} />
-                }
                 <Top>
                     {renderTop(G, ctx, ctx.currentPlayer === playerID, boardStates)}
                 </Top>
@@ -393,29 +177,17 @@ const Board = (props: Props) => {
                                 // is clicked card a valid target?
                                 if (cardInteraction.info.targets.find(s => s.cardID === cardID)) {
                                     console.log(`Clicked card is a valid target. Execute instruction with id <${cardInteraction.info.instructionID}>`);
+                                    const bsType = boardStates.find(s => s.info?.instructionID === cardInteraction.info.instructionID)?.type;
+                                    playExecuteDoSound(bsType);
                                     moves.executeDo(cardInteraction.info.instructionID, {
                                         protagonist: playerID, cardID
                                     });
                                     setCardInteraction(undefined);
-                                    moves.setUICardToCard(undefined)
                                 }
                             }
                         }}
-                        onStableCardMouseEnter={cardID => {
-                            if (cardInteraction?.key === "card_to_card") {
-                                moves.setUICardToCard({
-                                    sourceCardID: cardInteraction.info.sourceCardID,
-                                    instructionID: cardInteraction.info.instructionID,
-                                    targetCardID: cardID,
-                                    protagonist: playerID,
-                                })
-                            }
-                        }}
-                        onStableCardMouseLeave={cardID => {
-                            if (cardInteraction?.key === "card_to_card") {
-                                moves.setUICardToCard(undefined)
-                            }
-                        }}
+                        onStableCardMouseEnter={cardID => { }}
+                        onStableCardMouseLeave={cardID => { }}
                         onPlayerClick={plid => {
                             if (cardInteraction?.key === "card_to_player") {
                                 console.log("Detected click for cardInteraction with key <card_to_player>");
@@ -442,7 +214,7 @@ const Board = (props: Props) => {
                 </Main>
                 <Middle>
                     <DrawPileWrapper zIndexFocus={isHoveringOverHandCard}>
-                        {renderEndTurnButton(moves, playerID, () => { playEndTurnButtonSound() }, boardStates)}
+                        {renderEndTurnButton(moves, playerID, () => playEndTurnButtonSound(), boardStates)}
                         <MiddleLabel>Deck</MiddleLabel>
                         <DrawPile onClick={() => {
                             if (boardStates.find(s => s.type === "drawCard")) {
@@ -463,13 +235,10 @@ const Board = (props: Props) => {
                         }} isGlowing={boardStates.find(s => s.type === "drawCard" || s.type === "draw__clickOnDrawPile") !== undefined} count={G.drawPile.length} />
                     </DrawPileWrapper>
                     <AnimatePresence>
-
-                        {
-                            renderNeighLabel(G, ctx, moves, playerID)
-                        }
+                        <NeighPanel G={G} ctx={ctx} moves={moves} playerID={playerID} />
                     </AnimatePresence>
 
-                    {renderInfoLabel(G, ctx, playerID, boardStates)}
+                    <InfoPanel G={G} ctx={ctx} playerID={playerID} boardStates={boardStates} />
                     <MiddleLeftWrapper zIndexFocus={isHoveringOverHandCard}>
                         <div>
                             <MiddleLabel>Nursery</MiddleLabel>
@@ -486,245 +255,28 @@ const Board = (props: Props) => {
                     </MiddleLeftWrapper>
                 </Middle>
                 <Bottom>
-                    <Stable
-                        ref={stableRef}
-                        cards={[...G.stable[playerID], ...G.temporaryStable[playerID]].map(c => G.deck[c])}
-                        upgradeDowngradeCards={G.upgradeDowngradeStable[playerID].map(c => G.deck[c])}
-                        glowing={glowingCardIDs}
-                        highlightMode={stableHighlightMode}
-                        onStableItemClick={(evt, cardID) => {
-                            // initiate card to card interaction for destroy and steal actions
-                            if (cardInteraction === undefined) {
-                                // check if it is a destroy or steal action
-                                // check if the card that is clicked is the source for the action
-                                let boardState = _.first(boardStates.filter(s => s.info?.sourceCardID === cardID && (s.type === "destroy__cardToCard" || s.type === "steal__cardToCard" || s.type === "sacrifice__cardToCard" || s.type === "move__cardToCard" || s.type === "returnToHand__cardToCard" || s.type === "backKick__card_to_card" || s.type === "unicornswap1")));
-
-                                if (boardState) {
-                                    const cardRef = stableRef.current?.getStableItemRef(cardID)!;
-                                    const coord = cardRef.current!.getBoundingClientRect();
-                                    const from = {
-                                        x: coord.left + coord.width / 2.0,
-                                        y: coord.top + coord.height / 2.0,
-                                    };
-                                    setCardInteraction({
-                                        key: "card_to_card",
-                                        info: {
-                                            sourceCardID: cardID,
-                                            instructionID: boardState.info!.instructionID,
-                                            targets: boardState.info!.targets!,
-                                            currentMousePosition: { x: evt.clientX, y: evt.clientY },
-                                            startingMousePosition: { ...from }
-                                        }
-                                    });
-                                }
-
-                                boardState = _.first(boardStates.filter(s => (s.type === "swapHands__cardToPlayer" || s.type === "pullRandom__cardToPlayer" || s.type === "move2__cardToPlayer" || s.type === "makeSomeoneDiscard__cardToPlayer" || s.type === "unicornswap2" || s.type === "blatantThievery1") && s.info?.sourceCardID === cardID));
-                                if (boardState) {
-                                    const cardRef = stableRef.current?.getStableItemRef(cardID)!;
-                                    const coord = cardRef.current!.getBoundingClientRect();
-                                    const from = {
-                                        x: coord.left + coord.width / 2.0,
-                                        y: coord.top + coord.height / 2.0,
-                                    };
-                                    setCardInteraction({
-                                        key: "card_to_player",
-                                        info: {
-                                            sourceCardID: cardID,
-                                            instructionID: boardState.info!.instructionID,
-                                            targets: boardState.info!.targets!,
-                                            currentMousePosition: { x: evt.clientX, y: evt.clientY },
-                                            startingMousePosition: { ...from }
-                                        }
-                                    });
-                                }
-                            } else {
-                                // cancel card interaction 
-                                if ((cardInteraction.key === "card_to_card" || cardInteraction.key === "card_to_player") && cardInteraction.info.sourceCardID === cardID) {
-                                    setCardInteraction(undefined);
-                                    moves.setUICardToCard(undefined);
-                                    return;
-                                }
-
-                                if (cardInteraction?.key === "click_on_other_stable_card" || cardInteraction?.key === "card_to_card") {
-                                    console.log("Detected click for cardInteraction with key <click_on_other_stable_card>");
-                                    // is clicked card a valid target?
-                                    if (cardInteraction.info.targets.find(s => s.cardID === cardID)) {
-                                        console.log(`Clicked card is a valid target. Execute instruction with id <${cardInteraction.info.instructionID}>`);
-                                        moves.executeDo(cardInteraction.info.instructionID, {
-                                            protagonist: playerID, cardID
-                                        });
-                                        setCardInteraction(undefined);
-                                    }
-                                }
-                            }
-                        }}
-                        onStableItemMouseEnter={cardID => {
-                            playHubMouseOverSound();
-                            const o = openScenes.filter(([instr]) => instr.ui.info?.source === cardID);
-                            let targets: HoverTarget[] = [];
-                            o.forEach(([instruction]) => {
-                                targets = [...targets, ...findUITargets(G, ctx, instruction)];
-                            });
-
-                            // show hover targets if the hovered card is a source card
-                            if (o.length > 0) {
-                                setHoverTargets({ sourceCardID: cardID, targets });
-                            }
-                        }}
-                        onStableItemMouseLeave={cardID => {
-                            // if a card to card interaction is in progress we do not want to unhighlight the targets
-                            // thus we unhighlight the targets when are not in a card to card interaction
-                            // this if query checks this condition 
-                            if (cardInteraction?.key !== "card_to_card") {
-                                setHoverTargets(undefined);
-                            }
-                        }}
-                        renderAccessoryHoverItem={cardID => {
-                            let boardState = boardStates.find(s => s.type === "discard__popup__ask" && s.info?.sourceCardID === cardID);
-                            if (boardState) {
-                                const { instruction, scene } = _findInstruction(G, boardState.info!.instructionID)!;
-                                // this if condition is just for typescript interference
-                                if (instruction.do.key === "discard" && instruction.ui.type === "single_action_popup") {
-                                    return (
-                                        <CardPopupSingleAction
-                                            text={instruction.ui.info?.singleActionText!}
-                                            onClick={() => {
-                                                moves.commit(scene!.id);
-                                            }}
-                                        />
-                                    );
-                                }
-                            }
-
-                            boardState = boardStates.find(s => s.type === "bring__popup__ask" && s.info?.sourceCardID === cardID);
-                            if (boardState) {
-                                const { instruction, scene } = _findInstruction(G, boardState.info!.instructionID)!;
-                                // this if condition is just for typescript interference
-                                if (instruction.do.key === "bringToStable" && instruction.ui.type === "single_action_popup") {
-                                    return (
-                                        <CardPopupSingleAction
-                                            text={instruction.ui.info?.singleActionText!}
-                                            onClick={() => {
-                                                moves.commit(scene!.id);
-                                            }}
-                                        />
-                                    );
-                                }
-                            }
-
-                            boardState = boardStates.find(s => (s.type === "shakeUp" || s.type === "reset" || s.type === "shuffleDiscardPileIntoDrawPile") && s.info?.sourceCardID === cardID);
-                            if (boardState) {
-                                const { instruction } = _findInstruction(G, boardState.info!.instructionID)!;
-                                // this if condition is just for typescript interference
-                                if ((instruction.do.key === "shakeUp" || instruction.do.key === "reset" || instruction.do.key === "shuffleDiscardPileIntoDrawPile") && instruction.ui.type === "single_action_popup") {
-                                    return (
-                                        <CardPopupSingleAction
-                                            text={instruction.ui.info?.singleActionText!}
-                                            onClick={() => {
-                                                moves.executeDo(instruction.id, { protagonist: playerID, sourceCardID: cardID });
-                                            }}
-                                        />
-                                    );
-                                }
-                            }
-
-                            boardState = boardStates.find(s => (s.type === "revive") && s.info?.sourceCardID === cardID);
-                            if (boardState) {
-                                const { instruction } = _findInstruction(G, boardState.info!.instructionID)!;
-                                // this if condition is just for typescript interference
-                                if ((instruction.do.key === "revive") && instruction.ui.type === "single_action_popup") {
-                                    return (
-                                        <CardPopupSingleAction
-                                            text={instruction.ui.info?.singleActionText!}
-                                            onClick={() => {
-                                                setShowDiscardFinder(boardState?.info?.targets?.map((s: ReviveTarget) => ({ cardID: s.cardID })));
-                                            }}
-                                        />
-                                    );
-                                }
-                            }
-
-                            boardState = boardStates.find(s => (s.type === "reviveFromNursery") && s.info?.sourceCardID === cardID);
-                            if (boardState) {
-                                const { instruction } = _findInstruction(G, boardState.info!.instructionID)!;
-                                // this if condition is just for typescript interference
-                                if ((instruction.do.key === "reviveFromNursery") && instruction.ui.type === "single_action_popup") {
-                                    return (
-                                        <CardPopupSingleAction
-                                            text={instruction.ui.info?.singleActionText!}
-                                            onClick={() => {
-                                                setShowNurseryFinder(true);
-                                            }}
-                                        />
-                                    );
-                                }
-                            }
-
-                            boardState = boardStates.find(s => (s.type === "addFromDiscardPileToHand__single_action_popup") && s.info?.sourceCardID === cardID);
-                            if (boardState) {
-                                const { instruction } = _findInstruction(G, boardState.info!.instructionID)!;
-                                // this if condition is just for typescript interference
-                                if ((instruction.do.key === "addFromDiscardPileToHand") && instruction.ui.type === "single_action_popup") {
-                                    return (
-                                        <CardPopupSingleAction
-                                            text={instruction.ui.info?.singleActionText!}
-                                            onClick={() => {
-                                                setShowDiscardFinder(boardState?.info?.targets?.map((s: AddFromDiscardPileToHandTarget) => ({ cardID: s.cardID })));
-                                            }}
-                                        />
-                                    );
-                                }
-                            }
-
-                            boardState = boardStates.find(s => (s.type === "search__single_action_popup") && s.info?.sourceCardID === cardID);
-                            if (boardState) {
-                                const { instruction } = _findInstruction(G, boardState.info!.instructionID)!;
-                                // this if condition is just for typescript interference
-                                if ((instruction.do.key === "search") && instruction.ui.type === "single_action_popup") {
-                                    return (
-                                        <CardPopupSingleAction
-                                            text={instruction.ui.info?.singleActionText!}
-                                            onClick={() => {
-                                                setShowDeckFinder(boardState?.info?.targets);
-                                            }}
-                                        />
-                                    );
-                                }
-                            }
-
-
-
-                            return undefined;
-                        }}
-                        onPlaceHereClick={() => {
-                            if (cardInteraction?.key === "play_upgradeDowngradeCardFromHand__choose_target") {
-                                moves.playUpgradeDowngradeCard(playerID, playerID, cardInteraction.info.cardID);
-                                setCardInteraction(undefined);
-                            }
-                        }}
+                    <StableSection
+                        G={G}
+                        ctx={ctx}
+                        playerID={playerID}
+                        moves={moves}
+                        boardStates={boardStates}
+                        stableRef={stableRef}
+                        cardInteraction={cardInteraction}
+                        setCardInteraction={setCardInteraction}
+                        hoverTargets={hoverTargets}
+                        setHoverTargets={setHoverTargets}
+                        openScenes={openScenes}
+                        glowingCardIDs={glowingCardIDs}
+                        stableHighlightMode={stableHighlightMode}
+                        setShowDeckFinder={setShowDeckFinder}
+                        setShowDiscardFinder={setShowDiscardFinder}
+                        setShowNurseryFinder={setShowNurseryFinder}
+                        isHoveringOverHandCard={isHoveringOverHandCard}
+                        setHoveringOverHandCard={setHoveringOverHandCard}
+                        playHubMouseOverSound={playHubMouseOverSound}
+                        playExecuteDoSound={playExecuteDoSound}
                     />
-                    {renderHand(G, ctx, moves, playerID, boardStates, cardInteraction, setCardInteraction,
-                        (index) => {
-                            setHoveringOverHandCard(true);
-
-                            if (playerID === ctx.currentPlayer) {
-                                moves.setUIHoverHandIndex(index)
-                            }
-                        },
-                        (index) => {
-                            setHoveringOverHandCard(false);
-
-                            if (playerID === ctx.currentPlayer) {
-                                moves.setUIHoverHandIndex(undefined);
-                            }
-                        },
-                        () => {
-                            setHoveringOverHandCard(false)
-
-                            if (playerID === ctx.currentPlayer) {
-                                moves.setUIHoverHandIndex(undefined);
-                            }
-                        }, openScenes)}
                 </Bottom>
             </Wrapper>
             </AnimateSharedLayout>
@@ -732,72 +284,6 @@ const Board = (props: Props) => {
 }
 
 ////////////////////////////////
-
-const renderNeighLabel = (G: UnstableUnicornsGame, ctx: Ctx, moves: any, playerID: PlayerID) => {
-    if (!G.neighDiscussion) {
-        return null;
-    }
-
-    const currentRound = _.last(G.neighDiscussion.rounds)!;
-    let role: NeighLabelRole = "original_initiator";
-    let newInitiatorName: string | undefined = undefined;
-    const originalInitiatorName = G.players[parseInt(G.neighDiscussion.protagonist)].name;
-
-    if (G.neighDiscussion.rounds.length > 1) {
-        const beforeRound = _.last(G.neighDiscussion.rounds, 2)[0];
-        newInitiatorName = G.players[parseInt(_.findKey(beforeRound.playerState, val => val.vote === "neigh")!)].name;
-    }
-
-    if (G.neighDiscussion.protagonist === playerID) {
-        if (G.neighDiscussion.rounds.length > 1) {
-            const beforeRound = _.last(G.neighDiscussion.rounds, 2)[0];
-            if (beforeRound.playerState[playerID].vote === "neigh") {
-                role = "original_initiator";
-            } else {
-                role = "original_initiator_can_counterneigh";
-            }
-        } else {
-            role = "original_initiator";
-        }
-    } else if (currentRound.playerState[playerID].vote === "undecided") {
-        role = "open";
-    } else if (currentRound.playerState[playerID].vote === "neigh") {
-        role = "did_neigh";
-    } else if (currentRound.playerState[playerID].vote === "no_neigh") {
-        if (G.neighDiscussion.rounds.length > 1) {
-            const beforeRound = _.last(G.neighDiscussion.rounds, 2)[0];
-            if (beforeRound.playerState[playerID].vote === "neigh") {
-                role = "new_initiator";
-            } else {
-                role = "did_not_neigh";
-            }
-        } else {
-            role = "did_not_neigh";
-        }
-    }
-
-    const onPlayNeighClick = () => {
-        const neighCardOnHand = G.hand[playerID].map(c => G.deck[c]).find(c => c.type === "neigh");
-        if (neighCardOnHand) {
-            moves.playNeigh(neighCardOnHand.id, playerID, G.neighDiscussion!.rounds.length - 1)
-        } else {
-            const superNeigh = G.hand[playerID].map(c => G.deck[c]).find(c => c.type === "super_neigh");
-            moves.playSuperNeigh(superNeigh!.id, playerID, G.neighDiscussion!.rounds.length - 1)
-        }
-    }
-
-    const onDontPlayNeighClick = () => {
-        moves.dontPlayNeigh(playerID, G.neighDiscussion!.rounds.length - 1);
-    }
-
-    const didVote = currentRound.playerState[playerID].vote !== "undecided";
-
-    return (
-        <NeighLabelWrapper>
-            <NeighLabel card={G.deck[G.neighDiscussion.cardID]} originalInitiatorName={originalInitiatorName} newInitiatorName={newInitiatorName} role={role} didVote={didVote} numberOfNeighedCards={G.neighDiscussion.rounds.filter(s => s.state === "neigh").length} showPlayNeighButton={G.hand[playerID].map(c => G.deck[c]).filter(c => c.type === "neigh" || c.type === "super_neigh").length > 0 && G.playerEffects[playerID].find(s => s.effect.key === "you_cannot_play_neigh") === undefined} onPlayNeighClick={onPlayNeighClick} onDontPlayNeighClick={onDontPlayNeighClick} />
-        </NeighLabelWrapper>
-    );
-}
 
 const renderTop = (G: UnstableUnicornsGame, ctx: Ctx, isCurrentPlayer: boolean, boardStates: BoardState[]) => {
     if (isCurrentPlayer) {
@@ -814,13 +300,12 @@ const renderTop = (G: UnstableUnicornsGame, ctx: Ctx, isCurrentPlayer: boolean, 
         return (
             <HiddenHand
                 count={G.hand[ctx.currentPlayer].length}
-                active={G.uiHoverHandIndex}
             />
         );
     }
 };
 
-const renderEndTurnButton = (moves: any, playerID: PlayerID, playEndTurnButtonSound: () => void, boardStates: BoardState[]) => {
+const renderEndTurnButton = (moves: Moves, playerID: PlayerID, playEndTurnButtonSound: () => void, boardStates: BoardState[]) => {
     if (boardStates.find(s => s.type === "endTurn")) {
         return (
             <EndTurnButtonWrapper>
@@ -830,184 +315,6 @@ const renderEndTurnButton = (moves: any, playerID: PlayerID, playEndTurnButtonSo
     }
 
     return undefined;
-}
-
-const renderHand = (G: UnstableUnicornsGame, ctx: Ctx, moves: any, playerID: PlayerID, boardStates: BoardState[], cardInteraction: CardInteraction | undefined, setCardInteraction: (interaction: CardInteraction) => void, onMouseEnterHandCard: (index: number) => void, onMouseLeaveHandCard: (index: number) => void, onHandCardClick: () => void, openScenes: Array<[Instruction, Scene]>) => {
-    let glowingCards: CardID[] = [];
-
-    if (boardStates.find(s => s.type === "playCard")) {
-        // current player may play cards from its hand
-        if (cardInteraction?.key === "play_upgradeDowngradeCardFromHand__choose_target") {
-            // no glowing cards
-        } else {
-            glowingCards = G.hand[playerID].map(c => [canPlayCard(G, ctx, playerID, c), c]).filter(s => s[0]).map(s => s[1]) as CardID[];
-        }
-    } else if (boardStates.find(s => s.type === "neigh__playNeigh")) {
-        if (G.playerEffects[playerID].find(s => s.effect.key === "you_cannot_play_neigh") === undefined) {
-            glowingCards = G.hand[playerID].map(c => G.deck[c]).filter(c => c.type === "neigh" || c.type === "super_neigh").map(c => c.id);
-        }
-    } else if (boardStates.find(s => s.type === "discard" || s.type === "discard__popup__committed")) {
-        const discardState = boardStates.find(s => s.type === "discard" || s.type === "discard__popup__committed")!;
-        glowingCards = discardState.info!.targets!.map((c: DiscardTarget) => c.handIndex).map((c: number) => G.hand[playerID][c])
-    } else if (boardStates.find(s => s.type === "bring__popup__committed")) {
-        const discardState = boardStates.find(s => s.type === "discard" || s.type === "bring__popup__committed")!;
-        glowingCards = discardState.info!.targets!.map((c: BringToStableTarget) => c.cardID);
-    }
-
-    const onClick = (evt: React.MouseEvent, cardID: CardID) => {
-
-        if (boardStates.find(s => s.type === "playCard")) {
-            const cardsOnHandThatCanBePlayed = G.hand[playerID].map(c => [canPlayCard(G, ctx, playerID, c), c]).filter(s => s[0]).map(s => s[1]) as CardID[];
-            if (cardsOnHandThatCanBePlayed.includes(cardID)) {
-                const card = G.deck[cardID];
-                if (card.type === "upgrade" || card.type === "downgrade") {
-                    setCardInteraction({
-                        key: "play_upgradeDowngradeCardFromHand__choose_target",
-                        info: {
-                            instructionID: "____________",
-                            currentMousePosition: { x: evt.clientX, y: evt.clientY },
-                            startingMousePosition: { x: evt.clientX, y: evt.clientY },
-                            cardID
-                        },
-                    });
-                } else {
-                    moves.playCard(playerID, cardID);
-                }
-            }
-            onHandCardClick();
-        } else if (boardStates.find(s => s.type === "neigh__playNeigh")) {
-            if (G.deck[cardID].type === "neigh" || G.deck[cardID].type === "super_neigh") {
-                if (G.playerEffects[playerID].find(s => s.effect.key === "you_cannot_play_neigh") === undefined) {
-                    moves.playNeigh(cardID, playerID, G.neighDiscussion!.rounds.length - 1);
-                }
-            }
-            onHandCardClick();
-        } else if (boardStates.find(s => s.type === "discard" || s.type === "discard__popup__committed")) {
-            const discardState = boardStates.find(s => s.type === "discard" || s.type === "discard__popup__committed")!;
-            if (discardState.info!.targets!.find((s: DiscardTarget) => G.deck[G.hand[playerID][s.handIndex]].id === cardID)) {
-                // if click on hand card that is discardable
-                moves.executeDo(discardState.info!.instructionID, { protagonist: playerID, cardID });
-                onHandCardClick();
-            }
-        } else if (boardStates.find(s => s.type === "bring__popup__committed")) {
-            const discardState = boardStates.find(s => s.type === "bring__popup__committed")!;
-            if (discardState.info!.targets!.find((s: BringToStableTarget) => s.cardID === cardID)) {
-                // if click on hand card that is discardable
-                moves.executeDo(discardState.info!.instructionID, { protagonist: playerID, cardID });
-                onHandCardClick();
-            }
-        }
-    }
-
-    return (
-        <Hand cards={G.hand[playerID].map(c => G.deck[c])} glowingCards={glowingCards} onClick={onClick} onMouseEnterHandCard={(idx) => onMouseEnterHandCard(idx)} onMouseLeaveHandCard={(idx) => onMouseLeaveHandCard(idx)} />
-    );
-}
-
-const renderInfoLabel = (G: UnstableUnicornsGame, ctx: Ctx, playerID: PlayerID, boardStates: BoardState[]) => {
-    let text: string | undefined = undefined;
-    const openScenes = _findOpenScenesWithProtagonist(G, playerID);
-    const scenesInProgress = _findInProgressScenesWithProtagonist(G, playerID);
-
-    if (ctx.currentPlayer === playerID && ctx.activePlayers![playerID] === "beginning" && openScenes.length > 0) {
-        text = "One of your cards has an effect that can be activated. You can activate it and after that draw a card. You may also skip the effect and just draw a card."
-    }
-
-    if (ctx.currentPlayer === playerID && ctx.activePlayers![playerID] === "beginning" && scenesInProgress.length > 0) {
-        text = "One of your cards has an effect that must be activated. You must first activate it before you can draw a card."
-    }
-
-    if (playerID === ctx.currentPlayer) {
-        if (G.countPlayedCardsInActionPhase === 0 && G.neighDiscussion === undefined && ctx.activePlayers![playerID] === "action_phase") {
-            // action phase and no card has been played or drawn
-            // player may draw a card or play a card
-            text = "You can play a card from your hand or you can draw a card. You cannot do both."
-        }
-    }
-
-    if (boardStates.find(o => o.type === "discard" || o.type === "discard__popup__committed")) {
-        text = "Click on a card in your hand to discard that card."
-    }
-
-    if (boardStates.find(o => o.type === "destroy__click_on_card_in_stable")) {
-        text = "Click on a card in a player's stable to destroy that card."
-    }
-
-    if (boardStates.find(o => o.type === "sacrifice__clickOnCardInStable")) {
-        text = "Click on a card in your stable to sacrifice that card."
-    }
-
-    if (boardStates.find(o => o.type === "steal__cardToCard")) {
-        const boardState = boardStates.find(o => o.type === "steal__cardToCard")!;
-        const card = G.deck[boardState.info!.sourceCardID!];
-        text = `Click on ${card.title} and then click on another player's card which you'd like to steal.`
-    }
-
-    if (boardStates.find(o => o.type === "draw__clickOnDrawPile")) {
-        const boardState = boardStates.find(o => o.type === "draw__clickOnDrawPile");
-        text = `Click on the draw pile on the right to draw ${boardState?.info?.count} card.`
-    }
-
-    if (boardStates.find(o => o.type === "swapHands__cardToPlayer")) {
-        const boardState = boardStates.find(o => o.type === "swapHands__cardToPlayer")!;
-        const card = G.deck[boardState.info!.sourceCardID!];
-        text = `Click on ${card.title} and then on a player's name to trade hands with them.`
-    }
-
-    if (boardStates.find(o => o.type === "move__cardToCard")) {
-        const boardState = boardStates.find(o => o.type === "move__cardToCard")!;
-        const card = G.deck[boardState.info!.sourceCardID!];
-        text = `Click on ${card.title} and then on another card to move that card.`
-    }
-
-    if (boardStates.find(o => o.type === "move2__cardToPlayer")) {
-        const boardState = boardStates.find(o => o.type === "move2__cardToPlayer")!;
-        const card = G.deck[boardState.info!.sourceCardID!];
-        text = `Click on ${card.title} and then on any player's name to move the previously selected card into that player's stable.`
-    }
-
-    if (boardStates.find(o => o.type === "wait_for_other_players")) {
-        text = "Wait for other players...";
-    }
-
-    if (boardStates.find(o => o.type === "unicornswap1")) {
-        const boardState = boardStates.find(o => o.type === "unicornswap1")!;
-        const card = G.deck[boardState.info!.sourceCardID!];
-        text = `Click on ${card.title} and then on one of your cards which should be moved.`
-    }
-
-    if (boardStates.find(o => o.type === "unicornswap2")) {
-        const boardState = boardStates.find(o => o.type === "unicornswap2")!;
-        const card = G.deck[boardState.info!.sourceCardID!];
-        text = `Click on ${card.title} and then click on another player to which you would like to move the card.`
-    }
-
-    if (boardStates.find(o => o.type === "blatantThievery1")) {
-        const boardState = boardStates.find(o => o.type === "blatantThievery1")!;
-        const card = G.deck[boardState.info!.sourceCardID!];
-        text = `Click on ${card.title} and then click on a player to take a look at the player's hand.`
-    }
-
-    if (boardStates.find(o => o.type === "pullRandom__cardToPlayer")) {
-        const boardState = boardStates.find(o => o.type === "pullRandom__cardToPlayer")!;
-        const card = G.deck[boardState.info!.sourceCardID!];
-        text = `Click on ${card.title} and then click on a player to pull a random card from that player.`
-    }
-
-
-
-
-    if (!text) {
-        return undefined;
-    }
-
-    return (
-        <InfoLabelWrapper>
-            <InfoLabel>
-                {text}
-            </InfoLabel>
-        </InfoLabelWrapper>
-    );
 }
 
 ////////////////////////////////
@@ -1025,7 +332,7 @@ const Wrapper = styled(motion.div)`
 `;
 
 const Top = styled.div`
-    position: absolute; 
+    position: absolute;
     top: 0;
     z-index: 2;
     height: 50px;
@@ -1045,7 +352,7 @@ const Middle = styled.div`
 `;
 
 const MiddleLabel = styled.div`
-    color: white; 
+    color: white;
     margin-bottom: 0.4em;
 `;
 
@@ -1080,26 +387,10 @@ const Bottom = styled.div`
 `;
 
 const EndTurnButtonWrapper = styled.div`
-    position: absolute; 
-    top: 110%; 
-    z-index: 3000; 
+    position: absolute;
+    top: 110%;
+    z-index: 3000;
     width: 200px;
-`;
-
-const NeighLabelWrapper = styled.div`
-    width: 100%; 
-    display: flex;
-    justify-content: center;
-    position: absolute; 
-    top: -16px;
-`;
-
-const InfoLabelWrapper = styled.div`
-    width: 100%; 
-    display: flex;
-    justify-content: center;
-    position: absolute; 
-    top: -16px;
 `;
 
 export default Board;
