+++ src/game/operations/index.ts
@@ -1,0 +1,13 @@
+export * from './enter';
+export * from './steal';
+export * from './discard';
+export * from './destroy';
+export * from './sacrifice';
+export * from './draw';
+export * from './search';
+export * from './revive';
+export * from './move';
+export * from './swap';
+export * from './misc';
+export * from './executeDo';
+
